requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Migrations
alembic==1.13.1

# Auth
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ML
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.0.1
joblib==1.4.2

# Env
python-dotenv==1.0.1